Move Chrome PID helpers into chrome
Constraint: PID files, orphan reaping, and process guardians are Chrome lifecycle responsibilities, not domain-neutral utilities.\nRejected: Moving parent-watcher in the same PR | it watches the MCP host parent process and should be classified separately from Chrome child cleanup.\nRejected: Keeping pid-manager re-exported from src/utils/index.ts | that would preserve the utility bucket as a hidden domain API.\nConfidence: high\nScope-risk: moderate\nDirective: Import Chrome PID and child process guardian helpers from src/chrome, not src/utils.\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/chrome/pid-manager.test.ts tests/chrome/chrome-pid-manager.test.ts tests/chrome/process-guardian.test.ts tests/chrome/launcher-zombie.test.ts tests/headed/headed-fallback-manager.test.ts tests/tools/reap-orphans.test.ts tests/utils/sync-shutdown.test.ts tests/integration/embedded-server.test.ts\nTested: npm run build\nTested: npm run lint\nTested: npm run lint:tier\nTested: npm run lint:repo-structure\nTested: git diff --check\nNot-tested: Full Jest suite.
