+++ docs/releases/v1.10.1.md
@@ -171,7 +171,7 @@ release line):
   propagation path (socket close → `req.on('close')` → controller abort) has
   enough headroom on loaded macOS Actions runners. The test still fails fast
   if the propagation genuinely never occurs.
-- `tests/utils/process-guardian.test.ts` raises `waitForDead`'s default
+- `tests/chrome/process-guardian.test.ts` raises `waitForDead`'s default
   timeout from 4s to 15s, with each test given a 20s jest timeout. Windows
   CI spends several seconds on detached Node cold-start for the guardian
   plus the `taskkill /T /F /PID` round-trip before the child actually exits,
