+++ src/chrome/launcher.ts
@@ -7,8 +7,8 @@ import * as path from 'path';
 import * as fs from 'fs';
 import * as os from 'os';
 import { getGlobalConfig } from '../config/global';
-import { writeChromePid, removeChromePid, getChromePidFilePath, killProcessTree } from '../utils/pid-manager';
-import { spawnProcessGuardian } from '../utils/process-guardian';
+import { writeChromePid, removeChromePid, getChromePidFilePath, killProcessTree } from './pid-manager';
+import { spawnProcessGuardian } from './process-guardian';
 import { DEFAULT_VIEWPORT, DEFAULT_CHROME_LAUNCH_TIMEOUT_MS, DEFAULT_RESTORE_LAST_SESSION } from '../config/defaults';
 import type { WindowBoundsConfig } from '../config/window-bounds';
 import { getHeadedWindowArgs } from './launcher-window-args';
