+++ src/chrome/process-guardian.ts
