+++ src/index.ts
@@ -18,7 +18,7 @@ import { resolveHeadlessMode } from './config/headless-resolver';
 import { resolveCapabilityFilterOptions } from './config/capability-filter';
 import { resolveWindowBoundsConfig } from './config/window-bounds';
 import { ToolTier } from './config/tool-tiers';
-import { writePidFile, cleanOrphanedChromeProcesses } from './utils/pid-manager';
+import { writePidFile, cleanOrphanedChromeProcesses } from './chrome/pid-manager';
 import { installParentWatcher, ParentWatcherHandle } from './utils/parent-watcher';
 import { installIdleTimeout, IdleTimeoutHandle, parseDuration } from './utils/idle-timeout';
 import { getIdleState } from './utils/idle-state';
