+++ src/utils/index.ts
@@ -4,4 +4,3 @@
 
 export * from './atomic-file';
 export * from './json-validator';
-export * from './pid-manager';
