+++ src/utils/sync-shutdown.ts
@@ -17,7 +17,7 @@
  * Best-effort: we never throw; failures are logged to stderr.
  */
 
-import { killProcessTree } from './pid-manager';
+import { killProcessTree } from '../chrome/pid-manager';
 import { removeMarker } from '../chrome/ownership-marker';
 import { shouldKillChromeOnExit } from './session-resume-token';
