+++ tests/utils/sync-shutdown.test.ts
@@ -6,15 +6,15 @@ import {
   _resetForTesting,
 } from '../../src/utils/sync-shutdown';
 
-jest.mock('../../src/utils/pid-manager', () => {
-  const actual = jest.requireActual('../../src/utils/pid-manager');
+jest.mock('../../src/chrome/pid-manager', () => {
+  const actual = jest.requireActual('../../src/chrome/pid-manager');
   return {
     ...actual,
     killProcessTree: jest.fn(),
   };
 });
 
-import { killProcessTree } from '../../src/utils/pid-manager';
+import { killProcessTree } from '../../src/chrome/pid-manager';
 
 describe('sync-shutdown (#661 Phase 2)', () => {
   beforeEach(() => {
